personal_data/requirements.txt
bcrypt==4.3.0
mysql-connector-python==9.4.0